move transfer command to archiver.transfer
